WIP Change to celery task
Moved code to separate tasks.py file to run bulk update as celery task.
Currently is able to update for small courses but worker is killed for
large courses because of memory consumption.
